pyproject.toml
[project]
name = "hf-daily-papers-tg"
version = "0.1.0"
description = "Telegram bot that sends daily papers from Hugging Face."
authors = [
    { name = "Roman Solomatin", email = "risolomatin@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "huggingface-hub>=1.1.5",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-telegram-bot>=22.5",
]

[dependency-groups]
lint = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.6",
    "typos>=1.39.2",
]
tests = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-coverage>=0.0",
    "vcrpy>=7.0.0",
]
type = [
    "mypy>=1.18.2",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore= [
    "COM812",
    "D",
    "G",
]

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]


[tool.mypy]
python_version = "3.10"
strict = true
warn_redundant_casts = true
# align with mypy 2.0 release
warn_unreachable = true
local_partial_types = true
plugins = [
    "pydantic.mypy",
]
mypy_path = "src/"
disable_error_code = ["override"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
